Seven-file conflict resolution is union merges throughout: telegram's
fan-out, ordering seam and binding lookup keep their new structure while
errorprone's nullability discipline threads through them; CheckoutService
keeps the decomposed orchestrator, with the collaborators' annotation debt
settled by the now-promoted compiler in this same merge.

+++ platform/src/main/java/uz/horecaos/platform/web/cache/InProcessRateLimiter.java
@@ -151,7 +151,7 @@ public class InProcessRateLimiter implements RateLimiter {
         double elapsedSeconds = Duration.between(bucket.lastRefill(), now).toNanos() / 1_000_000_000.0;
         double windowSeconds = Math.max(0.001, policy.window().toMillis() / 1000.0);
         double perSecond = policy.permits() / windowSeconds;
-        return Math.min(policy.permits(), bucket.tokens() + elapsedSeconds * perSecond);
+        return Math.min((double) policy.permits(), bucket.tokens() + elapsedSeconds * perSecond);
     }
 
     private static Duration retryAfter(Policy policy) {
